Create README.md
EXISTING SYSTEM
Now prevailing system is in such a way that a user can book a mode of transportation by 
their own by the help of an app but individual booking is done there. Here the customer 
cannot use the privilege of sharing. They can mention the destination to reach and departure 
place but cannot share this and hence cannot invite other travelers.

PROPOSED SYSTEM 
This paper presents an android application that works collectively using Google APIs and 
maps. This is a ride-sharing app. Users with the same destination will be able to share rides 
with others who are of the same place to reach. The application will calculate their fares. 
Hence by this we can reduce our fare by not paying the actual amount.
Share My Ride
College of Applied Science, Thamarassery 6


MODULE DESCRIPTION

Admin
● Login
● Manage user
● Add emergency number
● View complaints and post reply
● View and manage registered vehicle

User
● Register
● Login
● Register vehicle
● Add and update route
● View available user 
● Book user
● View booking request and manage
● View booking status and manage
● Add payment
● View payment 
● Post complaint and view reply
● View emergency number and post alert
● Chat with user
● Post rating
● View rating